% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{calc_mean_BC}
\alias{calc_mean_BC}
\title{Calculated mean of normalised BC isotopes}
\usage{
calc_mean_BC(..., n_bc, q)
}
\arguments{
\item{...}{Channels to normalise and average.}

\item{q}{Percentile for normalisation.}
}
\value{
A vector.
}
\description{
Calculated mean of normalised BC isotopes
}
