% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{calc_mean_DNA}
\alias{calc_mean_DNA}
\title{Calculated mean of normalised Iridium isotopes}
\usage{
calc_mean_DNA(data, dna_channels, q)
}
\arguments{
\item{data}{SingleCellExperiment object with counts in linear scale}

\item{dna_channels}{Vector specifying the names of the DNA channels}

\item{q}{Quantile for normalisation.}
}
\value{
The SingleCellExperiment object with an extra column "mean_BC" in the counts assay.#'
}
\description{
Calculated mean of normalised Iridium isotopes
}
