% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{calc_mean_ir}
\alias{calc_mean_ir}
\title{Calculated scaled mean of Iridium isotopes}
\usage{
calc_mean_ir(..., q)
}
\arguments{
\item{...}{Channels to average.}

\item{q}{Quantile for normalisation.}
}
\value{
A vector.
}
\description{
Calculated scaled mean of Iridium isotopes
}
