% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{heatmap_compare_corr}
\alias{heatmap_compare_corr}
\title{Plot pearson correlation coefficients between markers on a double triangular heatmap (lower triangle: before RUCova, upper triangle: after RUCova).}
\usage{
heatmap_compare_corr(lower, upper)
}
\arguments{
\item{corr_values_before}{Matrix with pearson correlation coefficient between markers before RUCova.}

\item{corr_values_after}{Matrix with pearson correlation coefficient between markers after RUCova.}
}
\value{
#A heatmap
}
\description{
Plot pearson correlation coefficients between markers on a double triangular heatmap (lower triangle: before RUCova, upper triangle: after RUCova).
}
\examples{
plot_corr_before_after(corr_values_before,corr_values_after)
}
