% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HNSCC_data.R
\docType{data}
\name{HNSCC_data}
\alias{HNSCC_data}
\title{HNSCC data set}
\format{
A data frame with 108649 rows and 59 variables:
\describe{
\item{dose}{0Gy and 10Gy}
\item{line}{"Cal27","Cal33","UPCISCC099","UPCISCC131","UTSCC16A","UDSCC2","UPCISCC154","VUSCC147"  }
...
}
}
\usage{
data(HNSCC_data)
}
\description{
Is a tibble containing mass cytometry data of single-cell marker signals (rows = cells, columns = markers and metadata) in linear scale.
This data set should be clean, meaning you excluded beads, debris, doublets, dead cells, and single-cells are demultiplexed (important if you want to adapt the linear fits to the samples).
In this example we offer a mass cytometry data set consisting of 8 Head-and-Neck Squamous Cell Carcinoma (HNSCC) lines in irradiated (10 Gy) and control (0 Gy) conditions (Figure 2 and Figure 3 in the manuscript).
}
\keyword{datasets}
