% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_plots.R
\name{plot_pearson_corr}
\alias{plot_pearson_corr}
\title{Regress surrogates of unwanted covariance}
\usage{
plot_pearson_corr(data, surrogates, markers, approach, out_ruc_obj, clust)
}
\arguments{
\item{data}{A tibble.}

\item{surrogates}{Vector of surrogates to use for normalisation.}

\item{markers}{Vector of marker values to normalise.}

\item{approach}{Should be "per_sample" if XXXX}

\item{out_ruc_obj}{Some object}

\item{clust}{Some clust}
}
\value{
A ggplot object that plots
}
\description{
Regress surrogates of unwanted covariance
}
