% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{calc_mean_BC}
\alias{calc_mean_BC}
\title{Calculated mean of normalised highest BC per cell}
\usage{
calc_mean_BC(data, bc_channels, n_bc, q)
}
\arguments{
\item{data}{SingleCellExperiment object with counts in linear scale}

\item{bc_channels}{Vector specifying the names of the BC channels}

\item{n_bc}{number of barcoding isotopes per cell.}

\item{q}{Quantile for normalisation.}
}
\value{
The SingleCellExperiment object with an extra column "mean_BC" in the counts assay.
}
\description{
Calculated mean of normalised highest BC per cell
}
