% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{calc_mean_BC}
\alias{calc_mean_BC}
\title{Calculated mean of normalised highest BC per cell}
\usage{
calc_mean_BC(..., n_bc, q)
}
\arguments{
\item{...}{Channels to average.}

\item{n_bc}{number of barcoding isotopes per cell}

\item{q}{Quantile for normalisation.}
}
\value{
A vector.
}
\description{
Calculated mean of normalised highest BC per cell
}
