% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R, R/useful_functions_old.R
\name{calc_mean_DNA}
\alias{calc_mean_DNA}
\title{Calculated mean of normalised Iridium isotopes}
\usage{
calc_mean_DNA(..., q)

calc_mean_DNA(..., q)
}
\arguments{
\item{...}{Channels to average in linear scale. Asinh transformation is applied within the function.}

\item{q}{Quantile for normalisation.}

\item{sce}{SingleCellExperiment object with counts in linear scale}

\item{dna_channels}{Vector specifying the names of the DNA channels}
}
\value{
A vector.

A vector.
}
\description{
Calculated mean of normalised Iridium isotopes

Calculated mean of normalised Iridium isotopes
}
\examples{
data <- RUCova::HNSCC_data
data |> dplyr::mutate(mean_DNA = RUCova::calc_mean_DNA(DNA_191Ir, DNA_193Ir, q = 0.95))

}
