% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{calc_mean_DNA}
\alias{calc_mean_DNA}
\title{Calculated mean of normalised Iridium isotopes}
\usage{
calc_mean_DNA(..., q)
}
\arguments{
\item{...}{Channels to average in linear scale. Asinh transformation is applied within the function.}

\item{q}{Quantile for normalisation.}
}
\value{
A vector.
}
\description{
Calculated mean of normalised Iridium isotopes
}
\examples{
library(dplyr) 
data <- RUCova::HNSCC_data
data |> mutate(mean_DNA = RUCova::calc_mean_DNA(DNA_191Ir, DNA_193Ir, q = 0.95))

}
