% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{heatmap_compare_corr}
\alias{heatmap_compare_corr}
\title{Plot pearson correlation coefficients between markers on a double triangular heatmap (lower triangle: before RUCova, upper triangle: after RUCova).}
\usage{
heatmap_compare_corr(lower, upper)
}
\arguments{
\item{lower}{Matrix with pearson correlation coefficient between markers eg.: before RUCova, to be plotted in the lower triangle.}

\item{upper}{Matrix with pearson correlation coefficient between markers eg.: after RUCova, to be plotted in the upper triangle.}
}
\value{
#A heatmap
}
\description{
Plot pearson correlation coefficients between markers on a double triangular heatmap (lower triangle: before RUCova, upper triangle: after RUCova).
}
\examples{
\dontrun{
library(dplyr)
lower <-  data |> 
mutate_at(vars(m,x), asinh) |> 
select(m,x) |> 
cor(method= "pearson")
upper <-  data_reg |> 
mutate_at(vars(m,x), asinh) |> 
select(m,x) |> 
cor(method= "pearson")
heatmap_compare_corr(lower,upper)
}
}
