% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{heatmap_compare_corr}
\alias{heatmap_compare_corr}
\title{Plot pearson correlation coefficients between markers on a double triangular heatmap (lower triangle: before RUCova, upper triangle: after RUCova). If RUCova has not been applied, the output is a symmetric heatmap.}
\usage{
heatmap_compare_corr(
  sce,
  name_assay_before = "counts",
  name_assay_after = NULL,
  name_reduced_dim = NULL
)
}
\arguments{
\item{sce}{A SingleCellExperiment object with markers and SUCs in linear scale stored in the assay "name_assay". Asinh transformation is applied within the function.}

\item{name_assay_before}{A string specifying the name of the assay before RUCova (with original counts in linear scale).}

\item{name_assay_after}{A string specifying the name of the assay before RUCova (with original counts in linear scale).}

\item{name_reduced_dim}{A string specifying the name of the dimensionality reduction data stored under \code{reducedDim()}. If "PCA", then PCs will be included in the heatmao.}
}
\value{
#A heatmap with pearson correlation coefficients.
}
\description{
Plot pearson correlation coefficients between markers on a double triangular heatmap (lower triangle: before RUCova, upper triangle: after RUCova). If RUCova has not been applied, the output is a symmetric heatmap.
}
