% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucova.R
\name{rucova}
\alias{rucova}
\title{Remove unwanted covariance}
\usage{
rucova(
  data,
  markers,
  surrogates,
  apply_asinh_SUC,
  col_name_sample = NULL,
  center_surr = "per_sample",
  model = "simple",
  keep_offset = TRUE
)
}
\arguments{
\item{data}{A tibble.}

\item{markers}{Vector of marker values to normalise.}

\item{surrogates}{Vector of surrogates to use for normalisation.}
}
\value{
Normalised tibble.
}
\description{
Remove unwanted covariance
}
