% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucova.R
\name{rucova}
\alias{rucova}
\title{Remove unwanted covariance}
\usage{
rucova(
  sce,
  name_assay_before = "counts",
  markers,
  SUCs = c("mean_DNA", "mean_BC", "total_ERK", "pan_Akt"),
  name_reduced_dim = "PCA",
  apply_asinh_SUCs = TRUE,
  model = "interaction",
  col_name_sample = "line",
  center_SUCs = "across_samples",
  keep_offset = TRUE,
  name_assay_after = "counts_rucova"
)
}
\arguments{
\item{sce}{A SingleCellExperiment object with markers and SUCs in linear scale stored in the assay "name_assay_before". Asinh transformation is applied within the function.}

\item{name_assay_before}{A string specifying the name of the assay before RUCova (with original counts in linear scale).}

\item{markers}{Vector of marker names to normalise, y (in linear scale).}

\item{SUCs}{Vector of surrogates of unwanted covariance to use for normalisation, x (in linear scale).}

\item{name_reduced_dim}{string specifying the name of the dimensionality reduction result in the SingleCellExperiment sce.}

\item{apply_asinh_SUCs}{Apply (TRUE) or not (FALSE) asinh transformation to the SUCs. TRUE if SUCs are the measured surrogates, FALSE if SUCs are PCs.}

\item{model}{A character: "simple", "offset" or "interaction" defining the model.}

\item{col_name_sample}{A character indicating the column name in "data" defining each sample.}

\item{center_SUCs}{A character "across_samples" or "per_sample" defining how to center the SUCs in zero.}

\item{keep_offset}{Keep (TRUE) or not (FALSE) the offset intercept between samples.+}

\item{name_assay_after}{A string specifying the name of the assay after RUCova (with regressed counts in linear scale).}
}
\value{
The input SingleCellExperiment object with an additional assay (name_assay_after) and a list in the metadata containing all the model details.
}
\description{
Remove unwanted covariance
}
