% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucova.R
\name{rucova}
\alias{rucova}
\title{Remove unwanted covariance}
\usage{
rucova(
  data,
  markers,
  SUCs,
  apply_asinh_SUCs,
  col_name_sample = "line",
  center_SUCs = "across_samples",
  model = "interaction",
  keep_offset = TRUE
)
}
\arguments{
\item{data}{A tibble.}

\item{markers}{Vector of marker values to normalise, y.}

\item{SUCs}{Vector of surrogates of unwanted covariance to use for normalisation, x.}
}
\value{
Normalised tibble.
}
\description{
Remove unwanted covariance
}
