% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucova.R
\name{rucova}
\alias{rucova}
\title{Remove unwanted covariance}
\usage{
rucova(
  data,
  markers,
  SUCs,
  apply_asinh_SUCs,
  col_name_sample = "line",
  center_SUCs = "across_samples",
  model = "interaction",
  keep_offset = TRUE
)
}
\arguments{
\item{data}{A tibble with markers and SUCs in linear scale. Asinh transformation is applied within the function.}

\item{markers}{Vector of marker values to normalise, y (in linear scale).}

\item{SUCs}{Vector of surrogates of unwanted covariance to use for normalisation, x (in linear scale).}

\item{apply_asinh_SUCs}{Apply (TRUE) or not (FALSE) asinh transformation to the SUCs. TRUE if SUCs are the measured surrogates, FALSE if SUCs are PCs.}

\item{col_name_sample}{A character indicating the column name in "data" defining each sample.}

\item{center_SUCs}{A character "across_samples" or "per_sample" defining how to center the SUCs in zero.}

\item{model}{A character: "simple", "offset" or "interaction" defining the model.}

\item{keep_offset}{Keep (TRUE) or not (FALSE) the offset intercept between samples.}
}
\value{
Normalised tibble with marker and surrogate values in linear scale (as the input).
}
\description{
Remove unwanted covariance
}
\examples{
RUCova::rucova(data, markers_to_norm,SUCs = c("PC1","PC2","PC3","PC4"), apply_asinh_SUCs = FALSE, col_name_sample = "line", center_SUCs = "across_samples", model = "interaction", keep_offset = TRUE)
RUCova::rucova(data, markers_to_norm,SUCs = c("mean_DNA", "mean_BC", "total_ERK", "pan_Akt"), apply_asinh_SUCs = TRUE, col_name_sample = "line", center_SUCs = "across_samples", model = "interaction", keep_offset = TRUE)
}
